nids-project/raspberry_pi/requirements.txt
# ============================================================
#  NIDS Raspberry Pi — Python Requirements
#  Install with:  pip3 install -r requirements.txt
#
#  NOTE: tflite-runtime is ARM-optimised — do NOT install full
#        tensorflow on the Pi (too heavy).
# ============================================================

flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
eventlet>=0.33.0

paho-mqtt>=1.6.1

scapy>=2.5.0

numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
joblib>=1.1.0
pandas>=1.3.0

# TFLite runtime for Raspberry Pi (ARM)
# If the above fails use:  pip3 install --index-url https://google-coral.github.io/py-repo/ tflite_runtime
# tflite-runtime>=2.12.0

requests>=2.28.0
psutil>=5.9.0

# RPi.GPIO — only installs on actual Raspberry Pi hardware
# RPi.GPIO>=0.7.1
